create PydanticPackageManagerListener class
