+++ resources/META-INF/plugin.xml
@@ -222,10 +222,9 @@
         <postStartupActivity implementation="com.koxudaxi.pydantic.PydanticInitializer" order="last"/>
         <localInspection language="Python" shortName="PydanticTypeCheckerInspection" suppressId="PydanticTypeChecker" bundle="messages.PyPsiBundle" displayName="Type checker compatible with Pydantic" groupKey="INSP.GROUP.python" enabledByDefault="false"  level="WARNING" implementationClass="com.koxudaxi.pydantic.PydanticTypeCheckerInspection"/>
     </extensions>
-<!--    <projectListeners>-->
-<!--        <listener class="com.koxudaxi.pydantic.PydanticInitializer$PydanticActiveSdkListener"-->
-<!--                  topic="com.jetbrains.python.packaging.PyPackageManager$Listener"/>-->
-<!--    </projectListeners>-->
+    <projectListeners>
+        <listener class="com.koxudaxi.pydantic.PydanticPackageListener" topic="com.jetbrains.python.packaging.PyPackageManager$Listener"></listener>
+    </projectListeners>
     <extensions defaultExtensionNs="Pythonid">
         <typeProvider implementation="com.koxudaxi.pydantic.PydanticTypeProvider"/>
         <inspectionExtension implementation="com.koxudaxi.pydantic.PydanticIgnoreInspection"/>
